Always fixing the uris
I implemented an option API_ABSOLUTE_URIS which is set to 1 by default because I couldn't manage to make the ruby Restfully gem to work correctly with relative uris.
By the way, relative uris are really relatives now, if you set API_ABSOLUTE_URIS to 0. Before that, it was a complete mess...

My problem with restfully is that:

For example, /resources returns a collection of resource objects, each having a "jobs" link, like that:

            {
               "rel" : "collection",
               "href" : "resources/2/jobs",
               "title" : "jobs"
            }

If I ask Restfully to follow the jobs, it queries for /resources/resources/2/jobs which doesn't exists.

With absolute uris, I get:

            {
               "rel" : "collection",
               "href" : "/oarapi/resources/2/jobs",
               "title" : "jobs"
            }

And restfully correctly follows the /oarapi/resources/2/jobs of course.