Disable the 'setup' Makefile target used when installing from sources.
This target is only used when installing from sources and it only set
the shell for the user oar. Currently it break the installation from the
sources and the installation process is going to change.

So I prefer that people installing from the snapshots sources. Use
manually the target 'setup' after installing the module.