[scheduler] Fix race condition: resource state
On G5K when the nodes are put to Absent in the job epilog then a race condition
can occur if the scheduler launches a job before the resources are turned into
Absent by the module node_change_state.pl.
So some jobs can be deleted because the nodes become Absent which is not
expected.