[database] fix schema version (2.5.2)
The schema version update was forgotten during OAR 2.5.2 release, while
the schema was indeed changed.
As a result, a bug is reported for OAR 2.5.2 and OAR 2.5.3 and a notice
was posted to the release notes on the web site.
A patch is now added to Debian package 2.5.3-3 to fix the version.

This commit fixes that bug. (the patch to the debian package will have
to  be removed in the next version: 2.5.4)

The issues will still prevail however, for oar-database creation done
with OAR 2.5.2 or 2.5.3 (tgz or RPM or Debian before 2.5.3-3), for which
the schema version will need a fix by the OAR installation administror.