[scheduler] drop Scheduler.pm -> move code to oar_meta_sched
Hand made cherry pick from the master branch:
- commit 5b59ebb383e95ac220e193f19d15da312542221f
- Pierre Neyron <pierre.neyron@imag.fr> (mar. 12 févr. 2013 22:52:40
CET)
- [scheduler] drop scheduler.pm and rework oar_meta_sched

Some more code was retrieved and push to oar_meta_sched, from the diff
between the last version of Scheduler.pm in the master branch before
it's removal (in the before mentioned commit) and the last version in
this branch (2.5).

Hope nothing is missing.