[accounting] Optimized and fixed (memory leak)
The accounting script could eat a lot of memory when (re)generating an accounting
table from a big jobs table and could crash a system.
By using a placeholder syntax and a DBI prepare function out of the main loop, we
fix the memory leak and increase the speed of the process.

Furthermore, we found a DB (a rare case!) where start_time could be equal to "1" leading into erroneous consumption count.