[drawgantt-svg] bugfix display in google chrome/chromium
In recent version of chromium (e.g. Version 31.0.1650.63 Built on Debian
7.2, running on Debian 7.4 (238485)), the 1st load of the page was
broken, with the svg frame not resized to fit the window (resizing
manualy the window fixed the display).
This seems to be a race condition at the load time. This commit adds a
timer of 100ms before the initial load of the svg content.

Also, this commit tries to not call reload_content twice at startup.