[drawgantt-svg] fix hierarchy display in the case it uses cpuset
If cpuset is used as a resource properties in the hierarchy, draw a
rectangle for every cpuset, even if the next one has the same value.
Unlike resource properties, cpuset are not unique per se, but loops for
every new network address. Also cpuset are likely to be at the finest
grain of the hierarchy, so this workaround should be just fine.