[api] bugfix: Added X_API_PATH_PREFIX as an apache environement variable
Actually, request headers seems to not be available anymore with apache 2.4
So, we convert the needed headers to variables. It was done for X_REMOTE_IDENT in the previous commit, but we need to do it also for X_API_PATH_PREFIX.
The API is backward compatible (still using headers on apache implementations that serves them to the cgi)
Also simplified the config.