[oarnodesetting] factorize sql queries and fix concurrency
Work on issues #60 and #61.

Several SQL queries on "resources" and "resource_log" are factorized to speed
up the execution (only when '-r' and/or '--sql' options are used).

A table LOCK has been put on the tables "resources" and "resource_logs" to
avoid the deadlock between "oarnodesetting" and "node_change_state" module
using Pg database.
I am not sure if this is the right way to deal with this deadlock because a
freeze or a SIGSTP signal on the oarnodesetting process can stop the whole
thing... but this was already the case with MySQL.