[reservations] debug fallback function
When the "gantt_jobs_predictions" and/or "gantt_jobs_resources" are manually
flushed (or a crash when OAR is feeding these tables) then a code is used to
re-calculate the resources to assign to a reservation.
There was a bug which that took only Absent resources as available for
reservations.