[all_in_one_scheduler] skip useless adv. reservation code for the besteffort queue
Adv reseravtions cannot be besteffort: they are forbidden by admission
rules.

Not sure if it could be enabled: it would at least need some changes to
the code to behave like batch jobs with the gantt (only add current adv.
reservation when in the besteffort queue), and maybe more.

As a result, the following stages are useless for the besteffort queue:
- processing of reservations with missing resources
- processing of new reservations

=> As a result, if not forbidden by admission rule, oarsub for adv.
reservation won't get validated, users will have to hit Ctrl-C.