[api]: Secured apache auth config by header example
The proposed filtering could not be sufficient to allow the access to the API to only some specified users. The X_REMOTE_IDENT header should be removed to be sure that the API only gets the environment variable of the same name, after filtering.
Another problem was that the header could be passed as X_REMOTE_IDENT or X-REMOTE-IDENT (perl query module seems to understand both), so the filtering was not working if the client passed a form or another. X-REMOTE_IDENT and X_REMOTE-IDENT have also been removed, just in case, even if the tests did not lead to erroneous auth with those forms.