[Scheduler] Add support for job type -t state=permissive
1/ with -t state=permissive, a batch job is scheduled regardless of
having
some resources in state Absent or Suspected (already the case for
advance reservations).

However, the job will never run (it is always rescheduled in the
future), unless its resources become Alive at some time...

... Unless it has also the cosystem or noop job type, in which case it
will start regardless of the state of the resources: Absent, Suspected
or even Dead.

(could be usefull to overcome the lack of fairsharing for advance
reservations)

2/ with -t state=permissive -t noop/cosystem, an adv. reservation will
start regardless of the Aliveness of the resources (without it, OAR
waits for resources to come back and then remove them after the timeout
if still not Alive.