[oarsh] add srun like feature to oarsh: exec on subset of cgroup
ex:
Create job with 2 CPUs.
> oarsub -l /host=2/cpu=2 ...

Get the cpuset ids of the first cpu
> oarprint cpu -P host,cpuset
node1 0,2,4,6
node1 1,3,5,7
node2 0,2,4,6
node2 1,3,5,7

Execute a commande on the first cpu of node1
> export OAR_USER_CPUSET=$(oarprint cpu -P host,cpuset | grep node1 | head -n1 | cut -d\  -f2)
> oarsh node1

Then on node1:
> cat /dev/cpuset/oar/user_jobid/cpuset.cpus
0-7
> cat /proc/self/cpuset
/oar/user_jobid/0,2,4,6
> cat /dev/cpuset/oar/user_jobid/0,2,4,6/cpuset.cpus
0,2,4,6

And works as well for GPUs: given a job on a host with 4 GPUs
> OAR_USER_GPUDEVICE=1,2 oarsh node1
only gpu 1 and 2 (/dev/nvidia1 and /dev/nvidia2) will show up in
nvidia-smi, not gpu 0 and 3.

Of course, OAR_USER_CPUSET and OAR_USER_GPUDEVICE cannot be used to gain
access to resources which are not part of the job (thanks to the cgroup
hierarchy)