[scheduler] fix race condition with adv. res. container jobs with missing resources
Bug was:
1- with a container adv resa job in the case where 1 resource is absent
at the time of starting the job
2- then, 1st inner jobs get in the running state while the container
itself is still waiting
3- because all resources are not alive, the container start time is
postponed (after some time, the absent resources will be removed from
the job)
4- the first inner jobs are running but not placed in the default gantt
as the container gantt structure is not yet created (container job
starts later, so it is handled later in the loop)
5- the scheduling in the container job does not take into account the
first inner jobs (which defaulted to gantt 0). This results in the later
inner jobs to overlap with the firsts.

Notes:
- container jobid > inner jobids is unexpected but possible (yes, the
user has to guess the container jobid to give the `-t inner=`)
- but anyway, with the fairsharing, container job may be scheduled after
the inner jobs, because jobid are not handled in fifo.