[oarapi] replace workdir by directory in OAR's rest API
When directory is set in an API query to create a job (to something
other than the user's home directory) the job will try to start from
this directory. However, in "script mode" (using `script` in place of
`command` in the JSON payload) `directory` is not enough to start the
job properly. The subscript, is written in the directory setted by
`workdir`, so both `directory` and `workdir` must been inquired.

By replacing `workdir` by `directory`, this allow to have only one
parameter to modify the launching directory of a job.

By doing this, the directory is not passed to oarsub with the
--directory option, instead a `cd` is done to the directory.

Fixes #180