[NodeChangeState] on prologue/epilogue error, only set the first node to suspected
The execution of the prologue/epilogue is on the first node, so just set
this one to suspected.

This is arguable, because prologue/epilogue could try to change all the
job's nodes. But changing all the nodes is more likely done by the job
resource manager, so should mostly not be the case.