[NodeChangeState] factorize the resources state changing
When resources are changing of state, three SQL requests are made to
update the OAR database (one update in the resources and resources_logs
tables, and one insert in the resources_logs table). Those requests were
perform for each resources that require an update.
If, for example, a lot of nodes were changing state, it was generating
overhead on the database. This is now fixed by making bulk SQL queries,
processing all the resources at a time.