[accounting] change the behavior of accounting to be more efficient
By adding the new state 'NA' (Not Applicable) for jobs.accounted, we can
have only the new jobs that have finished between two oaraccounting
launches. Before, jobs not accountable (e.g because of non default
ressources assigned) were kept in the state 'NO'. The SQL query was then
not efficient and had a high cost for the DBMS.