[oarexec] Fix systemd-run execution mode for passive jobs
Using a transient systemd service works fine for starting a command, but
when killing oarexec at the end of a job, the corresponding systemd
service is not stopped.  As a result, the command specified by the user
(usually "sleep") keeps running forever.

This is because the user process is not a child of systemd-run: it is run
directly by the init process.

Use the --scope mode of systemd-run to fix the problem: it is meant
exactly for this situation, ensuring that the process will run as a child
of the parent process.  Systemd-run will only ensure that the cgroups are
correct, which is what we want.

There are two side-effects to using the --scope mode:

- the CWD is not reset, but it's not a big issue in our case: we already
  reset it from the inline Perl script before starting the user command.

- env variables are not reset: the user process has access to some
  variables like OARDO_BECOME_USER, OARDIR, etc.  This was already the
  case before introducing systemd-run, so this is unlikely to cause any
  new issue.