[oarsub] better verification of advance reservations input
The regex used to validate and extract the time of and advance
reservation now match only if months, days, hours, minutes and seconds
are valid (e.g. a month can't be superior to 12). For days, it can't be
superior to 31. If a day is set to 31 for a month containing only 30
days, oarsub (and Perl) will put time to the first of next month.