[api] better returned messages and add an OAR_API env variable
We don't mention oarsub anymore in the messages returned by the API.
OAR_JOB_ID is now printed only in verbose mode if inferior or equal to
0.
An environment variable OAR_API is setted by the API to be used in
admission rules.
The returned http status code is now 400 instead of 500 when die() is
called in admission rules.