[oarsh/job_resources_manager/oar_resources_init] Use hwloc physical ids + harden output
Prefer physical ids rather than logical ids, because logical ids depends
on what cpu/core/pu is in the cgroup (always starts from 0).

When using `OAR_USER_CPUSET=... oarsh` in a job reserving only part of
the resources of a host, this is a problem because logical ids may not
be "absolute".